refactor: order, ordercomplete page
