+++ client/src/api/carts.js
@@ -11,15 +11,13 @@ export const fetchCarts = async () => {
 
 // cart 추가
 // 수량(quantity)은 언제나 1개
-export const addCart = async ({ id, selectedSize }) => axios.post(`${url}/${id}`, { selectedSize });
+export const addCart = async ({ id, selectedSize }) => {
+  axios.post(`${url}/${id}`, { selectedSize });
+};
 
 // cart 수량 변경
 export const changeQuantity = async ({ id, selectedSize, quantity }) => {
-  try {
-    return axios.patch(`${url}/${id}`, { selectedSize, quantity });
-  } catch (e) {
-    return e.message;
-  }
+  axios.patch(`${url}/${id}`, { selectedSize, quantity });
 };
 
 // cart 삭제
