+++ client/src/components/Carts/OrderHistory.jsx
@@ -1,36 +1,16 @@
 import { useNavigate } from 'react-router-dom';
-import { useQuery } from '@tanstack/react-query';
-import { Button, Stack, Title, Group, Text, useMantineColorScheme } from '@mantine/core';
-import styled from '@emotion/styled';
-import { cartsQuery } from '../../api/query';
+import { Stack, Title, Group, Text, useMantineColorScheme } from '@mantine/core';
 import { PATH } from '../../constants';
-
-const CustomButton = styled(Button)`
-  display: block;
-  margin-top: 2rem;
-  padding: 1.8rem 2.4rem;
-  border-radius: 30px;
-  height: 6rem;
-  color: #fff;
-  font-size: 1.6rem;
-  font-weight: 'bold';
-
-  :hover {
-    background-color: #228be6;
-  }
-`;
+import CustomButton from '../CustomButton';
+import { useCountCarts, useTotalPrice } from '../../hooks/carts';
 
 const OrderHistory = () => {
   const { colorScheme } = useMantineColorScheme();
 
   const navigate = useNavigate();
 
-  const { data: countCarts } = useQuery(cartsQuery({ select: carts => carts.length }));
-  const { data: totalPrice } = useQuery(
-    cartsQuery({
-      select: carts => carts.reduce((acc, cart) => acc + cart.quantity * cart.price, 0),
-    })
-  );
+  const countCarts = useCountCarts();
+  const totalPrice = useTotalPrice();
 
   return (
     <Stack w="33.33333%" px="0.8rem" py="0.8rem" mb="2.4rem" spacing={0}>
