+++ client/src/components/CustomButton.jsx
@@ -1,0 +1,13 @@
+import { Button } from '@mantine/core';
+import styled from '@emotion/styled';
+
+const CustomButton = styled(Button)`
+  margin-top: 2rem;
+  padding: 1.8rem 2.4rem;
+  border-radius: 30px;
+  height: 6rem;
+  font-size: 1.6rem;
+  font-weight: 'bold';
+`;
+
+export default CustomButton;
