+++ client/src/components/Order/index.js
@@ -1,0 +1,9 @@
+export { default as Payment } from './Payment';
+export { default as AddressInfo } from './AddressInfo';
+export { default as EditAddress } from './EditAddress';
+export { default as EditAddressItem } from './EditAddressItem';
+export { default as InputAddress } from './InputAddress';
+export { default as Coupons } from './Coupons';
+export { default as SelectPaymentMethod } from './SelectPaymentMethod';
+export { default as CartHistory } from './CartHistory';
+export { default as CartHistoryItemList } from './CartHistoryItemList';
