+++ client/src/components/OrderComplete/index.js
@@ -1,0 +1,2 @@
+export { default as OrderInfo } from './OrderInfo';
+export { default as OrderProducts } from './OrderProducts';
