+++ client/src/components/index.js
@@ -10,3 +10,4 @@ export { default as SadIcon } from './Icon';
 export { default as NoProduct } from './NoProduct';
 export { default as SizeButton } from './SizeButton';
 export { default as DarkMode } from './DarkMode';
+export { default as CustomButton } from './CustomButton';
