+++ client/src/constants/initField.js
@@ -1,0 +1,7 @@
+const INIT_FIELD = {
+  info: false,
+  edit: false,
+  input: false,
+};
+
+export default INIT_FIELD;
