+++ client/src/hooks/carts/index.js
@@ -1,3 +1,5 @@
 export { default as useAddCartMutation } from './useAddCartMutation';
 export { default as useChangeQuantityMutation } from './useChangeQuantityMutation';
 export { default as useRemoveCartMutation } from './useRemoveCartMutation';
+export { default as useCountCarts } from './useCountCarts';
+export { default as useTotalPrice } from './useTotalPrice';
