+++ client/src/hooks/carts/useCountCarts.js
@@ -1,0 +1,10 @@
+import { useQuery } from '@tanstack/react-query';
+import { cartsQuery } from '../../api/query';
+
+const useCountCarts = () => {
+  const { data: countCarts } = useQuery(cartsQuery({ select: carts => carts.length }));
+
+  return countCarts;
+};
+
+export default useCountCarts;
